Split JDK/Android authentication libraries from runtime
