+++ ClientRuntimes/Java/azure-android-client-authentication/build.gradle
@@ -42,7 +42,6 @@ repositories {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile 'com.android.support:appcompat-v7:23.0.1'
-    compile 'com.microsoft.rest:client-runtime:0.0.2-SNAPSHOT'
     compile 'com.microsoft.aad:adal:1.1.11'
     testCompile 'junit:junit:4.11'
     testCompile 'junit:junit-dep:4.11'
