+++ ClientRuntimes/Java/azure-client-authentication/build.gradle
@@ -31,7 +31,6 @@ repositories {
 }
 
 dependencies {
-    compile 'com.microsoft.rest:client-runtime:0.0.2-SNAPSHOT'
     compile 'com.microsoft.azure:adal4j:1.1.2'
     testCompile 'junit:junit:4.11'
     testCompile 'junit:junit-dep:4.11'
