+++ ClientRuntimes/Java/azure-client-authentication/src/main/java/com/microsoft/rest/credentials/ApplicationTokenCredentials.java
@@ -7,7 +7,12 @@
 
 package com.microsoft.rest.credentials;
 
-import com.squareup.okhttp.OkHttpClient;
+import com.microsoft.aad.adal4j.AuthenticationContext;
+import com.microsoft.aad.adal4j.AuthenticationResult;
+import com.microsoft.aad.adal4j.ClientCredential;
+
+import java.io.IOException;
+import java.util.concurrent.Executors;
 
 /**
  * Token based credentials for use with a REST Service Client.
@@ -80,7 +85,34 @@ public class ApplicationTokenCredentials extends TokenCredentials {
     }
 
     @Override
-    public void applyCredentialsFilter(OkHttpClient client) {
-        client.interceptors().add(new ApplicationTokenCredentialsInterceptor(this));
+    public String getToken() throws IOException {
+        if (token == null) {
+            token = acquireAccessToken();
+        }
+        return token;
+    }
+
+    @Override
+    public void refreshToken() throws IOException {
+        token = acquireAccessToken();
+    }
+
+    private String acquireAccessToken() throws IOException {
+        String authorityUrl = this.getEnvironment().getAuthenticationEndpoint() + this.getDomain();
+        AuthenticationContext context = new AuthenticationContext(authorityUrl, this.getEnvironment().isValidateAuthority(), Executors.newSingleThreadExecutor());
+        AuthenticationResult result;
+        try {
+            result = context.acquireToken(
+                    this.getEnvironment().getTokenAudience(),
+                    new ClientCredential(this.getClientId(), this.getSecret()),
+                    null).get();
+        } catch (Exception e) {
+            throw new IOException(e.getMessage(), e);
+        }
+        if (result != null && result.getAccessToken() != null) {
+            return result.getAccessToken();
+        } else {
+            throw new IOException("Failed to acquire access token");
+        }
     }
 }
